// src/generated/api/organization/models/organizationStatus.ts
/**
 * Generated by orval v7.13.2 🍺
 * Do not edit manually.
 * DCTD-UTC Commerce API
 * Contract for storefront and admin applications
 * OpenAPI spec version: 1.0.0
 */

export type OrganizationStatus = (typeof OrganizationStatus)[keyof typeof OrganizationStatus];

// eslint-disable-next-line @typescript-eslint/no-redeclare
export const OrganizationStatus = {
  ACTIVE: 'ACTIVE',
  INACTIVE: 'INACTIVE',
} as const;